Basic uplift of dcp9022cdw to EAPI6
This is a very basic uplift of the dcp9022cdw ebuild from EAPI5 to
EAPI6. This is necessary to fix warnings arising from EAPI5 not
being supported any longer.

Besides stepping up the EAPI version, this change removes the
empty src_prepare phase and thereby utilizes the default
implementation of this phase.