Fix printer detection in cupswrapper script
Fix printer detection in cupswrapper/brother_lpdwrapper_mfcl3770.
Without fix, the script fails to detect the printer name correctly and
instead generates the name 'optbrotherPrintersmfcl3770cdw'. This is
because the regular expressions the wrapper uses expects the symlink in
/usr/libexec/cups/filter/brother_lpdwrapper_mfcl3770cdw to have an
absolute target instead of a relative one.

Because there is likely to always be a scenario where the regular
expressions in brother_lpdwrapper_mfcl3770 don't work, the solution here
is to remove the regular expression logic and hardcode the printer name
to mfcl3770cdw.

Signed-off-by: Nicholas Vinson <nvinson234@gmail.com>